add: auth route, jwt, web auth route

+++ api/src/routes/memories.ts
@@ -13,8 +13,15 @@ const bodySchema = z.object({
 })
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories',  async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify();
+    })
+    
+    app.get('/memories',  async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -28,13 +35,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
         });
     })
     
-    app.get('/memories/:id',  async (request) => {
+    app.get('/memories/:id',  async (request, reply) => {
         const { id } = paramsSchema.parse(request.params);
         const memoryExists = await prisma.memory.findUniqueOrThrow({
             where: {
                 id
             }
         });
+
+        if (!memoryExists.isPublic && memoryExists.userId !== request.user.sub) {
+            return reply.status(401).send('Unauthorized')
+        }
 
         return memoryExists;
     })
@@ -46,17 +57,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '4907a9c8-9d69-4f09-abae-baa1b9b06119'
+                userId: request.user.sub,
             }
         })
 
         return memory;
     })
     
-    app.put('/memories/:id',  async (request) => {
+    app.put('/memories/:id',  async (request, reply) => {
         const { id } = paramsSchema.parse(request.params);
         const {content, isPublic, coverUrl} = bodySchema.parse(request.body)
-        const updatedMemory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send('Unauthorized')
+        }
+        
+        memory = await prisma.memory.update({
             where: {
                 id
             },
@@ -67,11 +88,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
         })
 
-        return updatedMemory
+        return memory
     })
     
-    app.delete('/memories/:id',  async (request) => {
+    app.delete('/memories/:id',  async (request, reply) => {
         const { id } = paramsSchema.parse(request.params);
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+        
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send('Unauthorized')
+        }
+        
         await prisma.memory.delete({
             where: {
                 id
